Added dynamic spawn locations
With new logic, seed determines where spawns will be. Each player is assigned a sector of the map to spawn in, and the highest value of one of the dimensions of the simplex noise3array determines which cell w/in sector gets the spawn. Note:  This is still not perfect, as admirals could still spawn next to each other in some cases, and in some configurations last row/and/or col will never contain a spawn.